Move auth route to folder
